Move sanitize code into its own file.  Copy sanitize code (and test) to nodejs
