+++ nodejs/lancedb/arrow.ts
@@ -33,6 +33,7 @@ import {
   Float32,
 } from "apache-arrow";
 import { type EmbeddingFunction } from "./embedding/embedding_function";
+import { sanitizeSchema } from "./sanitize";
 
 /** Data type accepted by NodeJS SDK */
 export type Data = Record<string, unknown>[] | ArrowTable;
@@ -208,13 +209,14 @@ export function makeArrowTable(
   }
 
   const opt = new MakeArrowTableOptions(options !== undefined ? options : {});
+  if (opt.schema !== undefined && opt.schema !== null) {
+    opt.schema = sanitizeSchema(opt.schema);
+  }
   const columns: Record<string, Vector> = {};
   // TODO: sample dataset to find missing columns
   // Prefer the field ordering of the schema, if present
   const columnNames =
-    options?.schema != null
-      ? (options?.schema?.names as string[])
-      : Object.keys(data[0]);
+    opt.schema != null ? (opt.schema.names as string[]) : Object.keys(data[0]);
   for (const colName of columnNames) {
     if (
       data.length !== 0 &&
@@ -381,6 +383,10 @@ async function applyEmbeddings<T>(
     return table;
   }
 
+  if (schema !== undefined && schema !== null) {
+    schema = sanitizeSchema(schema);
+  }
+
   // Convert from ArrowTable to Record<String, Vector>
   const colEntries = [...Array(table.numCols).keys()].map((_, idx) => {
     const name = table.schema.fields[idx].name;
@@ -510,6 +516,9 @@ export async function fromRecordsToBuffer<T>(
   embeddings?: EmbeddingFunction<T>,
   schema?: Schema,
 ): Promise<Buffer> {
+  if (schema !== undefined && schema !== null) {
+    schema = sanitizeSchema(schema);
+  }
   const table = await convertToTable(data, embeddings, { schema });
   const writer = RecordBatchFileWriter.writeAll(table);
   return Buffer.from(await writer.toUint8Array());
@@ -527,6 +536,9 @@ export async function fromRecordsToStreamBuffer<T>(
   embeddings?: EmbeddingFunction<T>,
   schema?: Schema,
 ): Promise<Buffer> {
+  if (schema !== undefined && schema !== null) {
+    schema = sanitizeSchema(schema);
+  }
   const table = await convertToTable(data, embeddings, { schema });
   const writer = RecordBatchStreamWriter.writeAll(table);
   return Buffer.from(await writer.toUint8Array());
@@ -545,6 +557,9 @@ export async function fromTableToBuffer<T>(
   embeddings?: EmbeddingFunction<T>,
   schema?: Schema,
 ): Promise<Buffer> {
+  if (schema !== undefined && schema !== null) {
+    schema = sanitizeSchema(schema);
+  }
   const tableWithEmbeddings = await applyEmbeddings(table, embeddings, schema);
   const writer = RecordBatchFileWriter.writeAll(tableWithEmbeddings);
   return Buffer.from(await writer.toUint8Array());
@@ -555,6 +570,9 @@ export async function fromDataToBuffer<T>(
   embeddings?: EmbeddingFunction<T>,
   schema?: Schema,
 ): Promise<Buffer> {
+  if (schema !== undefined && schema !== null) {
+    schema = sanitizeSchema(schema);
+  }
   if (data instanceof ArrowTable) {
     return fromTableToBuffer(data, embeddings, schema);
   } else {
@@ -612,5 +630,5 @@ function alignTable(table: ArrowTable, schema: Schema): ArrowTable {
 
 // Creates an empty Arrow Table
 export function createEmptyTable(schema: Schema): ArrowTable {
-  return new ArrowTable(schema);
+  return new ArrowTable(sanitizeSchema(schema));
 }
